refactor(sandbox): split workflows.rs god object into modules
